+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Please refer to [UPGRADING](UPGRADING.md) guide for upgrading to a major version
 ### Fixed
 
 - Fail clearly when an HTTP response header line is invalid
+- Treat empty request protocol versions as HTTP/1.1
 
 
 ## 7.10.2 - 2026-05-20